requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
psycopg2-binary==2.9.4
whitenoise==6.2.0